+++ apps/backyard/src/app/championship/create.tsx
@@ -1,13 +1,21 @@
-import { Controller, FieldValues, useForm } from 'react-hook-form';
-
-import { useRules } from '@/hooks/domain/use-rules';
-import { Championship } from '@/model/domain/championship';
-import TextField from '@/components/form/text-field';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import { useNavigate } from 'react-router-dom';
 
 import { SelectField } from 'ui';
+import { Championship } from 'lib';
+
+import TextField from '@/components/form/text-field';
+import Button from '@/components/button';
+
+import { useRules } from '@/hooks/domain/use-rules';
+import { useChampionships } from '@/hooks/domain/use-championships';
+import { notify } from '@/utils/notify';
 
 export default function ChampionshipCreateView() {
+  const { championships, createChampionship } = useChampionships();
   const { rules } = useRules();
+
+  const lastNr = championships.reduce((max, c) => (c.nr > max ? c.nr : max), 0);
 
   const {
     control,
@@ -19,13 +27,41 @@ export default function ChampionshipCreateView() {
   } = useForm<Championship>({
     defaultValues: {
       id: '',
-      title: '',
-      nr: 1,
-      rulesId: rules.at(-1)?.id,
+      name: '',
+      nr: lastNr + 1,
+      rulesId: rules[0].id,
+      published: false,
+      completed: false,
     },
   });
 
-  function saveChampionship(championship: Championship) {}
+  const navigate = useNavigate();
+
+  const handleTitleChange = () => {
+    if (!dirtyFields.id) {
+      const name = getValues('name');
+      const stdPattern = /^([HRWE]).*(\d{2})\/?(\d{2})$/;
+      const match = name.match(stdPattern);
+      if (match) {
+        const secondLetter = match[0].match(/[HR]/) ? 'r' : 'm';
+        setValue(
+          'id',
+          `${match[1].toLowerCase() + secondLetter}${match[2] + match[3]}`,
+          { shouldValidate: true }
+        );
+      }
+    }
+  };
+
+  const saveChampionship: SubmitHandler<Championship> = async (
+    championship
+  ) => {
+    await notify(
+      createChampionship(championship),
+      `${championship.name} angelegt.`
+    );
+    navigate('..');
+  };
 
   return (
     <div className="space-y-4">
@@ -34,8 +70,39 @@ export default function ChampionshipCreateView() {
         <div className="shadow rounded-md bg-white">
           <form onSubmit={handleSubmit(saveChampionship)} noValidate>
             <div className="space-y-4 p-4">
-              <TextField label="Bezeichnung" {...register('title')} />
-              <TextField label="Kennung" {...register('id')} />
+              <TextField
+                autoFocus
+                label="Bezeichnung"
+                placeholder="Hinrunde 2019/20 oder WM 2014"
+                error={errors.name?.message}
+                {...register('name', {
+                  required: 'Pflichtfeld',
+                  onBlur: handleTitleChange,
+                })}
+              />
+              <TextField
+                label="Kennung"
+                placeholder="Eindeutige Kennung"
+                error={errors.id?.message}
+                {...register('id', {
+                  required: 'Pflichtfeld',
+                  pattern: {
+                    value: /[a-z]{2}[0-9]{4}/,
+                    message:
+                      'Genau sechs Zeichen - zwei Kleinbuchstaben und dann vier Ziffern',
+                  },
+                  maxLength: {
+                    value: 6,
+                    message:
+                      'Genau sechs Zeichen - zwei Kleinbuchstaben und dann vier Ziffern',
+                  },
+                  validate: {
+                    uniqueSlug: (id) =>
+                      !championships.some((c) => c.id === id) ||
+                      'Turnier mit dieser Kennung existiert schon',
+                  },
+                })}
+              />
               <TextField
                 label="Nummer"
                 type="number"
@@ -48,6 +115,11 @@ export default function ChampionshipCreateView() {
                 name="rulesId"
                 options={rules}
               />
+            </div>
+            <div className="bg-gray-50 px-4 py-3 text-right sm:px-6 space-x-4">
+              <Button primary type="submit">
+                Speichern
+              </Button>
             </div>
           </form>
         </div>