+++ apps/backyard/src/hooks/domain/use-championships.ts
@@ -1,8 +1,7 @@
 import { useRecoilValue } from 'recoil';
+import { Championship, createEntity } from 'lib';
 
-import { createEntity } from '@/firebase/db/repository/create-entity';
 import { championshipsState } from '@/states/domain/championships-state';
-import { Championship } from '@/model/domain/championship';
 
 export function useChampionships() {
   const championships = useRecoilValue(championshipsState);