+++ apps/backyard/src/model/domain/championship.ts
@@ -1,9 +1,0 @@
-import { BaseModel } from '@/firebase/db/base-model';
-
-export interface Championship extends BaseModel {
-  title: string;
-  nr: number;
-  rulesId: string;
-  published: boolean;
-  completed: boolean;
-}