+++ apps/backyard/src/utils/notify.ts
@@ -1,7 +1,7 @@
 import toast from 'react-hot-toast';
 
 export function notify(promise: Promise<unknown>, successMsg: string) {
-  toast.promise(promise, {
+  return toast.promise(promise, {
     loading: 'Speichern ...',
     success: successMsg,
     error: 'Hoppla, das hat nicht geklappt.',